Move linkage_checker from os/mac to generic
