+++ Library/Homebrew/linkage_checker.rb
